requirements.txt
streamlit
Pillow
numpy
pandas
